backend/requirements.txt
Django
djangorestframework==3.16.0
pillow==11.2.1
requests==2.32.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
django-debug-toolbar
django-redis==3.1.6